app/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
python-jose==3.4.0
python-memcached==1.62
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
